gaining momentum
It’s been a good space for progress so far today. Instead of focusing
on downloading the right plugins to achieve a certain look, I’ve been
focusing on just using simple tools, namely, the pen tool, to achieve
the look that I’m going for. Math symbols have been added, and the
thick, warm look is being maintained.